修复时钟精度丢失

+++ serial_ctl.py
@@ -45,7 +45,7 @@ class SerialPort(object):
     def set_clock(self, clock):
         print(f'clock changed to {clock}')
 
-        clock = int(clock) * 1000
+        clock = float(clock) * 1000
         frequency_divider = 2.5e7 / clock
         frequency_divider_str = hex(round(frequency_divider))[2:].zfill(6)
         command = f"f{frequency_divider_str[:3]}0{frequency_divider_str[-3:]}"
